Reducing direct calls to getSession

+++ src/world/Action/ItemAction.cpp
@@ -94,7 +94,7 @@ void ItemAction::handleCompanionItem()
   auto player = getSourceChara()->getAsPlayer();
 
   player->unlockCompanion( m_itemAction->data().Calcu0Arg[ 0 ] );
-  player->dropInventoryItem( static_cast< Common::InventoryType >( m_itemSourceContainer ), m_itemSourceSlot );
+  player->dropInventoryItem( static_cast< Common::InventoryType >( m_itemSourceContainer ), static_cast< uint8_t >( m_itemSourceSlot ) );
 }
 
 void ItemAction::handleMountItem()
@@ -102,5 +102,5 @@ void ItemAction::handleMountItem()
   auto player = getSourceChara()->getAsPlayer();
 
   player->unlockMount( m_itemAction->data().Calcu0Arg[ 0 ] );
-  player->dropInventoryItem( static_cast< Common::InventoryType >( m_itemSourceContainer ), m_itemSourceSlot );
+  player->dropInventoryItem( static_cast< Common::InventoryType >( m_itemSourceContainer ), static_cast< uint8_t >( m_itemSourceSlot ) );
 }
